Reorged update_listbuilder to make troubleshooting easier.  Removed debug logging where it's not needed.
