+++ lib_listbuilder/fleet.py
@@ -2,7 +2,7 @@ import logging
 import logging.handlers
 
 _handler = logging.handlers.WatchedFileHandler("/var/log/shrimpbot/shrimp.log")
-logging.basicConfig(handlers=[_handler], level=logging.DEBUG)
+logging.basicConfig(handlers=[_handler], level=logging.INFO)
 
 import re
 import sqlite3
